Move option classes to a specific namespace
