+++ src/Applications/UUPDownload/DownloadRequest/Process.cs
@@ -31,6 +31,7 @@ using System.Threading.Tasks;
 using UnifiedUpdatePlatform.Services.Composition.Database;
 using UnifiedUpdatePlatform.Services.WindowsUpdate;
 using UnifiedUpdatePlatform.Services.WindowsUpdate.Targeting;
+using UUPDownload.Options;
 using UUPDownload.Downloading;
 
 namespace UUPDownload.DownloadRequest
